chore: migrations

+++ backend/migrations/20250815175514_init_user_houses.sql
@@ -1,5 +1,5 @@
 CREATE TABLE user_houses (
-    user_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
-    house_id INTEGER REFERENCES houses(id) ON DELETE CASCADE,
+    user_id BIGINT REFERENCES users(id) ON DELETE CASCADE,
+    house_id BIGINT REFERENCES houses(id) ON DELETE CASCADE,
     PRIMARY KEY (user_id, house_id)
 );
